+++ vaadin-platform-fusion-test/src/test/java/com/vaadin/platform/fusion/offline/OfflineIT.java
@@ -49,4 +49,9 @@ public class OfflineIT extends ChromeDeviceTest {
           setConnectionType(NetworkConnection.ConnectionType.ALL);
       }
   }
+
+  @Override
+  protected String getTestPath() {
+    return "/";
+  }
 }
